changes upload

+++ app/Http/Controllers/Admin/ThemeOptionController.php
@@ -21,6 +21,8 @@ class ThemeOptionController extends Controller
         $input['secondary_color'] = $request->secondary_color;
         $input['footer_color'] = $request->footer_color;
         $input['copyright_color'] = $request->copyright_color;
+        $input['code'] = $request->code;
+        $input['option'] = isset($request->option[0]) ? 1 : 0;
 
         if ($request->hasFile('footer_bg_image')) {
 
@@ -34,17 +36,21 @@ class ThemeOptionController extends Controller
     public function update(Request $request, $id)
     {
 
+
         $update = ThemeOption::find($id);
         $input['primary_color'] = $request->primary_color;
         $input['secondary_color'] = $request->secondary_color;
         $input['footer_color'] = $request->footer_color;
         $input['copyright_color'] = $request->copyright_color;
+        $input['code'] = $request->code;
+        $input['option'] = isset($request->option[0]) ? 1 : 0;
 
         if ($request->hasFile('footer_bg_image')) {
+           if(!empty($update->footer_bg_image)){
             unlink('uploads/' . $update->footer_bg_image);
+           }
             $input['footer_bg_image'] = $request->file('footer_bg_image')->store('themeoption', 'uploads');
         }
-
 
         $update->update($input);
         return back()->with('message', 'Data updated success');
